refactor: move CompareGraphNodePropertiesFunctor to system control
