+++ expertsystem/solving/properties.py
@@ -7,41 +7,12 @@ these properties.
 from collections import OrderedDict
 from copy import deepcopy
 from itertools import permutations
-from typing import Dict, List, Optional, Set, Type
-
-import attr
+from typing import Dict, List, Set
 
 from expertsystem.particles import ParticleCollection
 
 from .topology import StateTransitionGraph
-from .types import InteractionProperties, NodeQuantumNumber, ParticleWithSpin
-
-
-class CompareGraphNodePropertiesFunctor:
-    """Functor for comparing graph elements."""
-
-    def __init__(
-        self,
-        ignored_qn_list: Optional[Set[Type[NodeQuantumNumber]]] = None,
-    ) -> None:
-        self.__ignored_qn_list = ignored_qn_list if ignored_qn_list else set()
-
-    def __call__(
-        self,
-        node_props1: Dict[int, InteractionProperties],
-        node_props2: Dict[int, InteractionProperties],
-    ) -> bool:
-        for node_id, node_props in node_props1.items():
-            other_node_props = node_props2[node_id]
-            if attr.evolve(
-                node_props,
-                **{x.__name__: None for x in self.__ignored_qn_list},
-            ) != attr.evolve(
-                other_node_props,
-                **{x.__name__: None for x in self.__ignored_qn_list},
-            ):
-                return False
-        return True
+from .types import ParticleWithSpin
 
 
 def filter_particles(
